refactoring module_user
moving stuff to single files. No need to be that complex
