--- module_user/models/__init__.py
@@ -1,1 +1,0 @@
-from .User import User
