--- module_user/views/SignUpViewset.py
@@ -1,12 +1,0 @@
-from rest_framework import viewsets, mixins
-from rest_framework.permissions import AllowAny
-
-from module_user.models.User import User
-
-from module_user.serializers.UserSerializer import UserSerializer
-
-
-class SignUpViewset(viewsets.GenericViewSet, mixins.CreateModelMixin):
-    queryset = User.objects.all()
-    serializer_class = UserSerializer
-    permission_classes = [AllowAny]
